+++ example/lib/main.dart
@@ -84,7 +84,6 @@ class _MyHomePageState extends State<MyHomePage> {
                 PDFConverter pdfConverter = PDFConverter(
                   backMatterDelta: null,
                   frontMatterDelta: null,
-                  customConverters: [],
                   document: _quillController.document.toDelta(),
                   fallbacks: [...loader.allFonts()],
                   onRequestBoldFont: (String fontFamily) async {
@@ -102,7 +101,7 @@ class _MyHomePageState extends State<MyHomePage> {
                   onRequestFont: (String fontFamily) async {
                     return loader.getFontByName(fontFamily: fontFamily);
                   },
-                  params: params,
+                  pageFormat: params,
                 );
                 final pw.Document? document = await pdfConverter.createDocument();
                 if (document == null) {