+++ lib/core/constant/constants.dart
@@ -5,6 +5,7 @@ class Constant {
   static const String DEFAULT_FONT_FAMILY = 'Arial';
   //to encode markdown characters to avoid detection
   static const List<double> default_heading_size = <double>[37, 30, 24, 18, 12];
+  static final RegExp newLinesInsertions = RegExp(r'^¶+');
   static final RegExp IMAGE_LOCAL_STORAGE_PATH_PATTERN =
       RegExp(r'^((\/[a-zA-Z0-9-_]+)+|\/)$');
   static final RegExp IMAGE_FROM_NETWORK_URL = RegExp(